ENH: Fix the unbounded recursion of the broken refactor.
Supports *args for constraints, and allows pass-through of non-control-point arguments. Such arguments should be simple json-serializable, so they can be written and read from node parameters.

For example, to enable usage like:

    logic.setConstraint(MP, 'midpoint', A, B)
    logic.setDistance(T, 'distance', R, 100)

It may be possible to allow MarkupConstraintsLogic to accept a `JSONDecoder` or `JSONEncoder` to support more types in future work.